% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.r
\name{author_contributions}
\alias{author_contributions}
\title{Total commits per person}
\usage{
author_contributions(commits)
}
\arguments{
\item{commits}{(data frame) the commits you want to inspect}
}
\value{
a data frame of stats
}
\description{
\code{author_contributions} shows stats as the name suggests
}
\details{
This function shows how many commits each contributor has made
}
\examples{
author_contributions(commits)
}

