% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrepo.r
\name{clone_org_repos}
\alias{clone_org_repos}
\title{Clone repos in repo_data}
\usage{
clone_org_repos(org_data)
}
\arguments{
\item{org_data}{the return value of \code{get_org_repo_data}}
}
\value{
a list of references to cloned git repositories
}
\description{
\code{clone_org_repos} clones all repos in the given list and returns a list of references
}
\details{
This function create a temporary folder to store local clones of repos
}

