% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.r
\name{commits_per_week}
\alias{commits_per_week}
\title{Show commits per week per repo}
\usage{
commits_per_week(commits)
}
\arguments{
\item{commits}{(data frame) commits for one or more repos}
}
\value{
a data frame with aggegates about how many commits per week per repo
}
\description{
\code{commits_per_week(commits)}
}
\details{
This function does what it says :)
}
\examples{
commits_per_week(commits)
}

