% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrepo.r
\name{get_all_commit_data}
\alias{get_all_commit_data}
\title{Extract and combine git log for all repos in the list of references}
\usage{
get_all_commit_data(repos)
}
\arguments{
\item{a}{list of git2r repo references}
}
\value{
a data frame of commits and associated information
}
\description{
\code{get_all_commit_data(repos)} gets all commit data :)
}
\details{
This function combines all commit information from a list of repos into one data frame
}

