% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrepo.r
\name{get_org_repo_data}
\alias{get_org_repo_data}
\title{Get repository data for the organisation from github}
\usage{
get_org_repo_data(url, auth)
}
\arguments{
\item{url}{a string url}

\item{auth}{a list with two named fields list(user = "", pw = "")}
}
\value{
a data frame with git repo information
}
\description{
\code{get_org_repo_data} returns a data frame with repository information
}
\details{
This function performs one or more HTTP requests aagainst the github API v3.
}

