% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrepo.r
\name{git_log_to_df}
\alias{git_log_to_df}
\title{Extract the git log for a repo}
\usage{
git_log_to_df(repo)
}
\arguments{
\item{repo}{a git2r reference to a local repo}
}
\value{
a data frame with all commits and metadata
}
\description{
\code{git_log_to_df(repo)} extracts the commit log and other useful information
}
\details{
This function operates on a git2r repo refernce
}

