% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viz.r
\name{project_heatmap}
\alias{project_heatmap}
\title{A heatmap of relative commit activity per repo per week}
\usage{
project_heatmap(commits_per_project)
}
\arguments{
\item{commits_per_project,}{aggregated per project per period - in a data frame}
}
\value{
a ggplot object
}
\description{
\code{project_heatmap(repos)}
}
\details{
This creates a picture of repo activity for the organisation
}
\examples{
project_heatmap(projects_over_time(commits))
}

