% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.r
\name{projects_over_time}
\alias{projects_over_time}
\title{Show relative commit activity per week over time for all repos}
\usage{
projects_over_time(commits)
}
\arguments{
\item{(data}{frame) commits the commits of one or more repos}
}
\value{
a data frame with summary stats
}
\description{
\code{projects_over_time} returns summary stats
}
\details{
This function shows relative activity (% of most commits per week, a bit random but yeah)
}
\examples{
projects_over_time(commits)
}

