% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrepo.r
\name{remove_local_clones}
\alias{remove_local_clones}
\title{Remove cloned repos}
\usage{
remove_local_clones(repos)
}
\arguments{
\item{repos}{the return value of clone_org_repos - a list of references}
}
\value{
nothing
}
\description{
\code{remove_local_clones(repos)} deletes git repos from temporary folder
}
\details{
This function loops through all references to repos and deltes them on the file-sysem.
}

