% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.r
\name{repo_language}
\alias{repo_language}
\title{Show repo language stats}
\usage{
repo_language(org_repo_data)
}
\arguments{
\item{org_repo_data}{(data frame) github data about an organisation's repo}
}
\value{
a data frame of summary stats
}
\description{
\code{repo_language(org_repo_data)} return a summary of main languages of all repos
}
\details{
This function is based on information from github and is not 100% accurate
}

